backend/requirements.txt
asgiref==3.5.2
channels==3.0.5
Django==4.0.6
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0
channels_redis==3.4.0
PyJWT==2.4.0
django-cors-headers==3.11.0
Pillow==9.1.0 
redis==4.3.1
gunicorn==20.1.0
whitenoise==5.3.0
